+++ server/webpa.go
@@ -44,6 +44,12 @@ const (
 var (
 	// ErrorNoPrimaryAddress is the error returned when no primary address is specified in a WebPA instance
 	ErrorNoPrimaryAddress = errors.New("No primary address configured")
+
+	// strongCipherSuites are the tls.CipherSuite values that are safe for TLS versions less than 1.3
+	strongCipherSuites = []uint16{
+		tls.TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384,
+		tls.TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384,
+	}
 )
 
 // executor is an internal type used to start an HTTP server.  *http.Server implements
@@ -279,6 +285,9 @@ func (b *Basic) New(logger log.Logger, handler http.Handler) *http.Server {
 			Certificates: certs,
 			MinVersion:   b.minVersion(),
 			MaxVersion:   b.maxVersion(),
+
+			// ensure strong ciphers when the TLS version is 1.2 or less
+			CipherSuites: strongCipherSuites,
 		}
 
 		if len(b.ClientCACertFile) > 0 {